Handle non-JSON responses (#118)
* Only try to parse JSON for json-related content-types
* Specify that the client accepts application/json

Responses to attachment upload requests return a JSON body with the Content-Type header set to 'text/html'. If we sepcify that we accept 'application/json', then the server will return the proper Content-Type header. This required a re-run of VCRs.

From the docs:

Note that the Content-type of the server response can lie. This behavior, while not strictly correct according to the protocol, is intended. There is a bug in Microsoft Internet Explorer that will cause it to navigate away from a page executing a JavaScript client app at the end of an upload if the server returns the (correct) type of application/json. The server will return the correct Content-type, instead of text/html, if the client includes application/json in the request's Accept header.

Source: https://www.pivotaltracker.com/help/api/#Uploading_File_Attachments